+++ apps/switchBot.js
@@ -3,13 +3,10 @@ import { UCPlugin } from '../models/index.js'
 import _ from 'lodash'
 
 function init() {
-  if (!_.isArray(switchBotData)) return
   setTimeout(() => {
     const groupData = UCPr.defaultCfg.getConfig('group')
     tempData = _.cloneDeep(groupData)
-    for (const groupId of switchBotData) {
-      _.set(groupData, `${groupId}.enable`, ['UC-switchBot'])
-    }
+    switchBotData.forEach(groupId => _.set(groupData, `${groupId}.enable`, ['UC-switchBot']))
   }, 20)
 }
 
@@ -48,18 +45,15 @@ export default class UCSwitchBot extends UCPlugin {
   }
 
   init() {
-    if (!switchBotData) {
-      if (!Check.file(Path.switchBotjson)) {
-        file.JSONsaver(Path.switchBotjson, [987654321])
-      }
-      tempData = _.cloneDeep(UCPr.defaultCfg.getConfig('group'))
+    if (!Check.file(Path.switchBotjson)) {
+      switchBotData = file.JSONsaver(Path.switchBotjson, [])
+    } else {
       switchBotData = file.JSONreader(Path.switchBotjson)
-      Data.watch(Path.switchBotjson, () => {
-        switchBotData = file.JSONreader(Path.switchBotjson)
-      })
-      init()
-      Data.watch(Path.get('botConfig', 'group.yaml'), init)
     }
+    tempData = _.cloneDeep(UCPr.defaultCfg.getConfig('group'))
+    Data.watch(Path.switchBotjson, () => (switchBotData = file.JSONreader(Path.switchBotjson)))
+    init()
+    Data.watch(Path.get('botConfig', 'group.yaml'), init)
   }
 
   async openBot() {
@@ -76,17 +70,14 @@ export default class UCSwitchBot extends UCPlugin {
   open() {
     const oriData = _.get(tempData, `${this.groupId}.enable`)
     if (!oriData || _.isEqual(oriData, ['UC-switchBot'])) {
-      _.unset(this.groupData, `${this.groupId}.enable`)
-    } else {
-      _.set(this.groupData, `${this.groupId}.enable`, oriData)
+      return _.unset(this.groupData, `${this.groupId}.enable`)
     }
+    _.set(this.groupData, `${this.groupId}.enable`, oriData)
   }
 
   async closeBot() {
     if (!this.verifyPermission(this.Cfg.use)) return false
-    if (this.isClose) {
-      return this.reply('当前已经是关闭状态了哦~', true)
-    }
+    if (this.isClose) return this.reply('当前已经是关闭状态了哦~', true)
     _.set(this.groupData, `${this.groupId}.enable`, ['UC-switchBot'])
     switchBotData.push(this.groupId)
     file.JSONsaver(Path.switchBotjson, switchBotData)